requirements.txt
# Claude Tools Dependencies

# Core Google API integration
google-auth>=2.15.0
google-auth-oauthlib>=0.8.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.70.0

# Environment management
python-dotenv>=1.0.0

# Optional AI integrations
anthropic>=0.5.0
openai>=1.0.0

# Utilities
requests>=2.28.0
psutil>=5.9.0

# Development
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0